Add in locking and variable renames
There are now locks around the chats and user/online sets to avoid race conditions. Attributes of the request argument are also now put into variables when they're used multiple times. This is to mitigate the issues with invalid variables being referenced since the linter doesn't catch the protobuf cases.